+++ src/components/molecules/MenuCart/MenuCart.module.css
@@ -1,75 +1,3 @@
-.main { 
-  width: 35%;
-  position: absolute;
-  top: 0;
-  color: #1e1e1e;
-  justify-content: center;
-}
-
-.list_wrapper {
-  list-style: none;
-  margin-bottom: 35%;
-  padding: 10px;
-}
-
-.list_item {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  margin-bottom: 10px;
-}
-
-.item_thumbnail {
-  height: 80px;
-  width: 80px;
-  border-radius: 10px;
-  object-fit: cover;
-  margin-right: 10px;
-}
-
-.item_info_container {
-  width: 85%;
-  padding: 10px;
-}
-
-.item_name_del_container {
-  display: flex;
-  flex-direction: row;
-  align-items: space-between;
-  justify-content: space-between;
-}
-
-.trash_button {
-  background: none;
-  border: none;
-  cursor: pointer;
-  margin: 0;
-  padding: 0;
-}
-
-.trash_icn_red {
-  border-radius: 50%;
-}
-
-.quantity_container {
-  display: inline-flex;
-  background: #fff;
-  padding: 5px 8px;
-  border-radius: 20px;
-}
-
-.quantity_buttons {
-  background: #FEFEFE;
-  border: none;
-  border-radius: 50%;
-  cursor: pointer;
-}
-
-.quantity_wrapper {
-  margin-right: 5px;
-  margin-left: 5px;
-}
-
 @media (max-width: 900px){
   .main {
     width: 90%;
@@ -78,4 +6,12 @@
   .list_wrapper {
     margin-bottom: 45%;
   }
+}
+
+.main { 
+  width: 35%;
+  position: absolute;
+  top: 0;
+  color: #1e1e1e;
+  justify-content: center;
 }
