+++ src/components/organisms/CartItems/CartItems.module.css
@@ -1,0 +1,69 @@
+@media (max-width: 900px){
+   .list_wrapper {
+      margin-bottom: 45%;
+   }
+}
+
+.list_wrapper {
+   list-style: none;
+   margin-bottom: 35%;
+   padding: 10px;
+}
+
+.list_item {
+   display: flex;
+   align-items: center;
+   width: 100%;
+   margin-bottom: 10px;
+}
+
+.item_thumbnail {
+   height: 80px;
+   width: 80px;
+   border-radius: 10px;
+   object-fit: cover;
+   margin-right: 10px;
+}
+
+.item_info_container {
+   width: 85%;
+   padding: 10px;
+}
+
+.item_name_del_container {
+   display: flex;
+   flex-direction: row;
+   align-items: space-between;
+   justify-content: space-between;
+}
+
+.trash_button {
+   background: none;
+   border: none;
+   cursor: pointer;
+   margin: 0;
+   padding: 0;
+}
+
+.trash_icn_red {
+   border-radius: 50%;
+}
+
+.quantity_container {
+   display: inline-flex;
+   background: #fff;
+   padding: 5px 8px;
+   border-radius: 20px;
+}
+
+.quantity_buttons {
+   background: #FEFEFE;
+   border: none;
+   border-radius: 50%;
+   cursor: pointer;
+}
+
+.quantity_wrapper {
+   margin-right: 5px;
+   margin-left: 5px;
+}
